Simple refactoring App.js to different components
