+++ react-front-end/src/components/__test__/App.test.js
